Move settings files and some txUI files to features folder
