+++ apps/mobile-wallet/src/contexts/walletConnect/WalletConnectSessionProposalModal.tsx
@@ -40,9 +40,9 @@ import InfoBox from '~/components/InfoBox'
 import { ModalScreenTitle, ScreenSection } from '~/components/layout/Screen'
 import SpinnerModal from '~/components/SpinnerModal'
 import { ModalContent, ModalContentProps } from '~/features/modals/ModalContent'
+import { persistSettings } from '~/features/settings/settingsPersistentStorage'
 import usePersistAddressSettings from '~/hooks/layout/usePersistAddressSettings'
 import { useAppDispatch, useAppSelector } from '~/hooks/redux'
-import { persistSettings } from '~/persistent-storage/settings'
 import { initializeKeyringWithStoredWallet } from '~/persistent-storage/wallet'
 import { selectAddressesInGroup } from '~/store/addresses/addressesSelectors'
 import { newAddressGenerated, selectAllAddresses, syncLatestTransactions } from '~/store/addressesSlice'
