+++ apps/mobile-wallet/src/features/settings/EditWalletName.tsx
