+++ apps/mobile-wallet/src/features/settings/WalletDeleteModal.tsx
@@ -18,6 +18,7 @@ along with the library. If not, see <http://www.gnu.org/licenses/>.
 
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
+import { Platform } from 'react-native'
 
 import { sendAnalytics } from '~/analytics'
 import AppText from '~/components/AppText'
@@ -26,17 +27,19 @@ import Input from '~/components/inputs/Input'
 import { ModalScreenTitle, ScreenSection } from '~/components/layout/Screen'
 import SpinnerModal from '~/components/SpinnerModal'
 import { useWalletConnectContext } from '~/contexts/walletConnect/WalletConnectContext'
-import { ModalContent, ModalContentProps } from '~/features/modals/ModalContent'
+import BottomModal from '~/features/modals/BottomModal'
+import { ModalContent } from '~/features/modals/ModalContent'
+import withModalWrapper from '~/features/modals/withModalWrapper'
 import { useAppDispatch, useAppSelector } from '~/hooks/redux'
 import { deleteWallet } from '~/persistent-storage/wallet'
 import { walletDeleted } from '~/store/wallet/walletActions'
 import { showExceptionToast } from '~/utils/layout'
 
-interface WalletDeleteModalProps extends ModalContentProps {
+interface WalletDeleteModalProps {
   onDelete: () => void
 }
 
-const WalletDeleteModal = ({ onDelete, ...props }: WalletDeleteModalProps) => {
+const WalletDeleteModal = withModalWrapper<WalletDeleteModalProps>(({ id, onDelete }) => {
   const dispatch = useAppDispatch()
   const walletName = useAppSelector((s) => s.wallet.name)
   const { resetWalletConnectStorage } = useWalletConnectContext()
@@ -46,8 +49,6 @@ const WalletDeleteModal = ({ onDelete, ...props }: WalletDeleteModalProps) => {
   const [isLoading, setIsLoading] = useState(false)
 
   const handleDeleteConfirmPress = async () => {
-    props.onClose && props.onClose()
-
     setIsLoading(true)
 
     try {
@@ -66,38 +67,44 @@ const WalletDeleteModal = ({ onDelete, ...props }: WalletDeleteModalProps) => {
   }
 
   return (
-    <>
-      <ModalContent verticalGap {...props}>
-        <ScreenSection>
-          <ModalScreenTitle>⚠️ {t('Delete "{{ walletName }}"?', { walletName })}</ModalScreenTitle>
-        </ScreenSection>
-        <ScreenSection>
-          <AppText color="secondary" size={18}>
-            {t('Do you really want to delete this wallet from your device?')}
-          </AppText>
-          <AppText color="secondary" size={18}>
-            {t('You can always restore it later using your secret recovery phrase.')}
-          </AppText>
-          <AppText color="secondary" size={18}>
-            {t('If so, please enter the wallet name below, and hit the delete button.')}
-          </AppText>
-        </ScreenSection>
-        <ScreenSection>
-          <Input label={t('Wallet name')} value={inputWalletName} onChangeText={setInputWalletName} />
-        </ScreenSection>
-        <ScreenSection>
-          <Button
-            title={t('Delete')}
-            variant="alert"
-            onPress={handleDeleteConfirmPress}
-            disabled={inputWalletName !== walletName}
-            iconProps={{ name: 'trash' }}
-          />
-        </ScreenSection>
-      </ModalContent>
-      <SpinnerModal isActive={isLoading} />
-    </>
+    <BottomModal
+      id={id}
+      maximisedContent={Platform.OS === 'ios'}
+      Content={(props) => (
+        <>
+          <ModalContent verticalGap {...props}>
+            <ScreenSection>
+              <ModalScreenTitle>⚠️ {t('Delete "{{ walletName }}"?', { walletName })}</ModalScreenTitle>
+            </ScreenSection>
+            <ScreenSection>
+              <AppText color="secondary" size={18}>
+                {t('Do you really want to delete this wallet from your device?')}
+              </AppText>
+              <AppText color="secondary" size={18}>
+                {t('You can always restore it later using your secret recovery phrase.')}
+              </AppText>
+              <AppText color="secondary" size={18}>
+                {t('If so, please enter the wallet name below, and hit the delete button.')}
+              </AppText>
+            </ScreenSection>
+            <ScreenSection>
+              <Input label={t('Wallet name')} value={inputWalletName} onChangeText={setInputWalletName} />
+            </ScreenSection>
+            <ScreenSection>
+              <Button
+                title={t('Delete')}
+                variant="alert"
+                onPress={handleDeleteConfirmPress}
+                disabled={inputWalletName !== walletName}
+                iconProps={{ name: 'trash' }}
+              />
+            </ScreenSection>
+          </ModalContent>
+          <SpinnerModal isActive={isLoading} />
+        </>
+      )}
+    />
   )
-}
+})
 
 export default WalletDeleteModal
