+++ apps/mobile-wallet/src/screens/LoginWithPinScreen.tsx
@@ -28,7 +28,7 @@ import { useAppDispatch, useAppSelector } from '~/hooks/redux'
 import { useBiometrics } from '~/hooks/useBiometrics'
 import RootStackParamList from '~/navigation/rootStackRoutes'
 import { getStoredWalletMetadata, migrateDeprecatedMnemonic } from '~/persistent-storage/wallet'
-import { allBiometricsEnabled } from '~/store/settings/settingsActions'
+import { allBiometricsEnabled } from '~/features/settings/settingsActions'
 import { mnemonicMigrated, walletUnlocked } from '~/store/wallet/walletActions'
 import { showExceptionToast } from '~/utils/layout'
 import { resetNavigation } from '~/utils/navigation'
