+++ apps/mobile-wallet/src/store/store.ts
@@ -21,6 +21,7 @@ import { configureStore } from '@reduxjs/toolkit'
 
 import fundPasswordSlice from '~/features/fund-password/fundPasswordSlice'
 import modalSlice from '~/features/modals/modalSlice'
+import settingsSlice, { settingsListenerMiddleware } from '~/features/settings/settingsSlice'
 import addressDiscoverySlice from '~/store/addressDiscoverySlice'
 import contactsSlice from '~/store/addresses/contactsSlice'
 import addressesSlice from '~/store/addressesSlice'
@@ -28,7 +29,6 @@ import appSlice from '~/store/appSlice'
 import confirmedTransactionsSlice from '~/store/confirmedTransactionsSlice'
 import loadersSlice from '~/store/loadersSlice'
 import pendingTransactionsSlice from '~/store/pendingTransactionsSlice'
-import settingsSlice, { settingsListenerMiddleware } from '~/store/settingsSlice'
 import walletSlice from '~/store/wallet/walletSlice'
 import walletGenerationSlice from '~/store/walletGenerationSlice'
